+++ changelog.md
@@ -5,4 +5,7 @@
 - ua-parser-js - detect mobile devices
 - Social icons of Facebook and Instagram
 
+### Fixed
+- only activated users are allowed to login
+
 ## [1.0.0] - 2019-05-11